# apps/web/app/api/payments/unmatched/route.ts
/**
 * @swagger
 * /api/payments/unmatched:
 *   get:
 *     summary: List unmatched transactions with suggestions
 *     description: Returns up to 100 unmatched bank transactions, each enriched with suggested company matches based on tax ID exact match or name similarity (bigram Jaccard index).
 *     tags: [Payments]
 *     security:
 *       - cookieAuth: []
 *     responses:
 *       200:
 *         description: Array of unmatched transactions with suggested companies
 *         content:
 *           application/json:
 *             schema:
 *               type: array
 *               items:
 *                 type: object
 *                 properties:
 *                   id:
 *                     type: string
 *                     format: uuid
 *                   sender_name:
 *                     type: string
 *                   sender_inn:
 *                     type: string
 *                   amount:
 *                     type: number
 *                   entry_date:
 *                     type: string
 *                     format: date
 *                   suggested_companies:
 *                     type: array
 *                     items:
 *                       type: object
 *                       properties:
 *                         company_id:
 *                           type: string
 *                           format: uuid
 *                         company_name:
 *                           type: string
 *                         tax_id:
 *                           type: string
 *                           nullable: true
 *                         confidence:
 *                           type: number
 *                         reason:
 *                           type: string
 *       401:
 *         description: Authentication required
 *       403:
 *         description: Admin or dispatcher access required
 *       500:
 *         description: Internal server error
 */

export const dynamic = 'force-dynamic'

import { NextResponse } from 'next/server'
import { createServerClient } from '@/lib/supabase/server'
import { requireAdminOrDispatcher } from '@/middleware/auth'

export async function GET() {
  try {
    await requireAdminOrDispatcher()
    const supabase = createServerClient() as any

    // Get unmatched transactions
    const { data: unmatched, error: unmatchedError } = await supabase
      .from('bank_transactions')
      .select('*')
      .eq('status', 'unmatched')
      .order('entry_date', { ascending: false })
      .limit(100)

    if (unmatchedError) throw unmatchedError
    if (!unmatched || unmatched.length === 0) return NextResponse.json([])

    // Batch fetch: all companies at once (instead of per-transaction)
    const { data: allCompanies } = await supabase.from('companies').select('id, name, tax_id')

    const companies = allCompanies || []

    // Index companies by tax_id for O(1) lookup
    const companiesByTaxId = new Map<string, typeof companies>()
    for (const c of companies) {
      if (!c.tax_id) continue
      const existing = companiesByTaxId.get(c.tax_id) || []
      existing.push(c)
      companiesByTaxId.set(c.tax_id, existing)
    }

    // Match all transactions client-side (single pass, no extra queries)
    const results = unmatched.map((txn: any) => {
      const suggestions: Array<{
        company_id: string
        company_name: string
        tax_id: string | null
        confidence: number
        reason: string
      }> = []

      // Tax ID exact match
      if (txn.sender_inn) {
        const innMatches = companiesByTaxId.get(txn.sender_inn) || []
        for (const c of innMatches.slice(0, 3)) {
          suggestions.push({
            company_id: c.id,
            company_name: c.name,
            tax_id: c.tax_id,
            confidence: 1.0,
            reason: 'Tax ID exact match',
          })
        }
      }

      // Name similarity fallback
      if (suggestions.length === 0 && txn.sender_name) {
        const senderLower = txn.sender_name.toLowerCase()
        const scored = companies
          .map((c: any) => ({
            ...c,
            score: basicSimilarity(c.name.toLowerCase(), senderLower),
          }))
          .filter((c: any) => c.score > 0.2)
          .sort((a: any, b: any) => b.score - a.score)
          .slice(0, 3)

        for (const c of scored) {
          suggestions.push({
            company_id: c.id,
            company_name: c.name,
            tax_id: c.tax_id,
            confidence: c.score,
            reason: 'Name similarity',
          })
        }
      }

      return { ...txn, suggested_companies: suggestions }
    })

    return NextResponse.json(results)
  } catch (error: any) {
    console.error('Error fetching unmatched payments:', error)

    if (error.name === 'UnauthorizedError') {
      return NextResponse.json({ error: 'Authentication required' }, { status: 401 })
    }
    if (error.name === 'ForbiddenError') {
      return NextResponse.json({ error: 'Admin access required' }, { status: 403 })
    }
    return NextResponse.json({ error: 'Internal server error' }, { status: 500 })
  }
}

/**
 * Basic string similarity (Jaccard index on character bigrams)
 * Used as a rough client-side fallback
 */
function basicSimilarity(a: string, b: string): number {
  if (a === b) return 1
  if (a.length < 2 || b.length < 2) return 0

  const bigramsA = new Set<string>()
  const bigramsB = new Set<string>()

  for (let i = 0; i < a.length - 1; i++) bigramsA.add(a.slice(i, i + 2))
  for (let i = 0; i < b.length - 1; i++) bigramsB.add(b.slice(i, i + 2))

  let intersection = 0
  for (const bg of bigramsA) {
    if (bigramsB.has(bg)) intersection++
  }

  return intersection / (bigramsA.size + bigramsB.size - intersection)
}