# apps/web/src/lib/swagger-spec.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "RouteHub API",
    "version": "1.0.0",
    "description": "API documentation for RouteHub — fleet management, payments, documents, and inspections."
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development"
    },
    {
      "url": "https://geosafety.routehub.ge",
      "description": "Production (master)"
    },
    {
      "url": "https://team2.routehub.ge",
      "description": "Team 2"
    },
    {
      "url": "https://team3.routehub.ge",
      "description": "Team 3"
    }
  ],
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "sb-access-token",
        "description": "Supabase session cookie (set automatically after login)"
      },
      "cronSecret": {
        "type": "apiKey",
        "in": "header",
        "name": "authorization",
        "description": "Bearer token matching CRON_SECRET env var (for scheduled jobs)"
      }
    },
...
  },
  "security": [
    {
      "cookieAuth": []
    }
  ],
  "paths": {
...
    "/api/payments/unmatched": {
      "get": {
        "summary": "List unmatched transactions with suggestions",
        "description": "Returns up to 100 unmatched bank transactions, each enriched with suggested company matches based on tax ID exact match or name similarity (bigram Jaccard index).",
        "tags": ["Payments"],
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Array of unmatched transactions with suggested companies",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string",
                        "format": "uuid"
                      },
                      "sender_name": {
                        "type": "string"
                      },
                      "sender_inn": {
                        "type": "string"
                      },
                      "amount": {
                        "type": "number"
                      },
                      "entry_date": {
                        "type": "string",
                        "format": "date"
                      },
                      "suggested_companies": {
                        "type": "array",
                        "items": {
                          "type": "object",
                          "properties": {
                            "company_id": {
                              "type": "string",
                              "format": "uuid"
                            },
                            "company_name": {
                              "type": "string"
                            },
                            "tax_id": {
                              "type": "string",
                              "nullable": true
                            },
                            "confidence": {
                              "type": "number"
                            },
                            "reason": {
                              "type": "string"
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Admin or dispatcher access required"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
  },
...
}